% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_mfe_data.R
\name{calculate_fine_bd}
\alias{calculate_fine_bd}
\title{Calculate the bulk density of the fine fraction}
\usage{
calculate_fine_bd(df)
}
\arguments{
\item{df}{a data.frame read from the NSDR Viewer SQLite}
}
\value{
a data.frame with added/processed data
}
\description{
Calculate the bulk density of the fine fraction
}
