% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_mfe_data.R
\name{calculate_stocks}
\alias{calculate_stocks}
\title{Calculate Carbon and Nitrogen Stocks}
\usage{
calculate_stocks(df)
}
\arguments{
\item{df}{a data.frame read from the NSDR Viewer SQLite}
}
\value{
a data.frame with two additional columns for SOC and total N stocks
}
\description{
Calculate Carbon and Nitrogen Stocks
}
