% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orchestrate_export.R
\name{cms_export}
\alias{cms_export}
\title{Orchestrates the export of MfE data from a SQLite to a Excel spreadsheet}
\usage{
cms_export(sqlite_fn, xlsx_fn, process_data = FALSE)
}
\arguments{
\item{sqlite_fn}{Path to a SQLite file, typically downloaded from the NSDR Viewer app}

\item{xlsx_fn}{A file path to save the xlsx file}

\item{process_data}{Does the data needs post-processing (splining, adding NZSC, etc)? Currently not implemented and set to FALSE.}
}
\value{
Nothing, but writes Excel spreadsheet to disk
}
\description{
Orchestrates the export of MfE data from a SQLite to a Excel spreadsheet
}
\author{
Pierre Roudier
}
