% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mfe_sqlite.R
\name{cms_read}
\alias{cms_read}
\title{Read a SQLite DB into a data.frame}
\usage{
cms_read(fn, view = "MfE_Carbon_data", legacy = FALSE)
}
\arguments{
\item{fn}{Path to a SQLite file}

\item{view}{Name of the SQLite View to load. Defaults to "MfE_Carbon_data". Part of the legacy system mentioned below.}

\item{legacy}{Boolean, default to FALSE This is the old system relying purely
on Views. It is kept temporarily for testing purposes.}
}
\value{
a data.frame
}
\description{
Read a SQLite DB into a data.frame
}
\author{
Pierre Roudier
}
