% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mfe_xlsx.R
\name{cms_write}
\alias{cms_write}
\title{Writes data.frame to Excel file}
\usage{
cms_write(df, fn)
}
\arguments{
\item{df}{a data.frame of MfE soil carbon data}

\item{fn}{A file path to save the xlsx file}
}
\value{
Nothing, but writes Excel spreadsheet to disk
}
\description{
Writes data.frame to Excel file
}
\author{
Pierre Roudier
}
