% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_mfe_data.R
\name{process_mfe_data}
\alias{process_mfe_data}
\title{Process the Downloaded Data}
\usage{
process_mfe_data(df)
}
\arguments{
\item{df}{a data.frame read from the NSDR Viewer SQLite}
}
\value{
a data.frame with added/processed data
}
\description{
Process the Downloaded Data
}
