% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sqlite.R
\name{read_sqlite}
\alias{read_sqlite}
\title{Read a SQLite DB into a data.frame}
\usage{
read_sqlite(fn)
}
\arguments{
\item{fn}{Path to a SQLite file}
}
\value{
a data.frame
}
\description{
Read a SQLite DB into a data.frame
}
\author{
Pierre Roudier
}
