% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mfe_xlsx.R
\name{write_mfe_xlsx}
\alias{write_mfe_xlsx}
\title{Writes data.frame to Excel file}
\usage{
write_mfe_xlsx(df, fn)
}
\arguments{
\item{df}{a data.frame of MfE soil carbon data}

\item{fn}{A file path to save the xlsx file}
}
\value{
Nothing, but writes Excel spreadsheet to disk
}
\description{
Writes data.frame to Excel file
}
\author{
Pierre Roudier
}
